+++ tests/minimal.spec.tsx
@@ -1,6 +1,16 @@
 import { expect, test } from 'vitest';
-import { Router } from 'waku-navigation';
+import {
+  Pending,
+  Router,
+  Slice,
+  unstable_useNavigationStatus,
+  useRouter,
+} from 'waku-navigation';
 
 test('exports', () => {
   expect(Router).toBeDefined();
+  expect(Pending).toBeDefined();
+  expect(Slice).toBeDefined();
+  expect(useRouter).toBeDefined();
+  expect(unstable_useNavigationStatus).toBeDefined();
 });